Clean up the code for class b3dtile::Tile
